+++ scripts/installation_scripts/qq-metacentrum-install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Installs qq on your current desktop and on the computing nodes of all Metacentrum clusters.
-# Script version: 0.4.0
+# Script version: 0.5.0
 
 set -euo pipefail
 
@@ -13,7 +13,6 @@ QQ_VERSION="v__VERSION__"
 
 # GitHub release assets
 INSTALL_SCRIPT_URL="https://github.com/VachaLab/qq/releases/download/${QQ_VERSION}/qq-install.sh"
-INSTALL_SCRIPT_CUSTOM_RC_URL="https://github.com/VachaLab/qq/releases/download/${QQ_VERSION}/qq-install-custom-rc.sh"
 SETUP_SCRIPT_URL="https://github.com/VachaLab/qq/releases/download/${QQ_VERSION}/qq-set-scratch.sh"
 RELEASE_URL="https://github.com/VachaLab/qq/releases/download/${QQ_VERSION}/qq-release.tar.gz"
 
@@ -90,11 +89,10 @@ for HOME_DIR in "${TARGET_HOMES[@]}"; do
 done
 
 # install qq on computers with local home directories (samson)
-# samson opens login shell after ssh, we need to add path to qq into .bash_profile
 for HOST in "${LOCAL_HOME_HOSTS[@]}"; do
     echo "INFO    [qq metacentrum installer] Installing qq on ${HOST}..."
     if ssh -o BatchMode=yes -o ConnectTimeout=20 "${HOST}" "exit" 2>/dev/null; then
-        ssh "${HOST}" "curl -fsSL ${INSTALL_SCRIPT_CUSTOM_RC_URL} | bash -s -- \${HOME} .bash_profile ${RELEASE_URL}"
+        ssh "${HOST}" "curl -fsSL ${INSTALL_SCRIPT_URL} | bash -s -- \${HOME} ${RELEASE_URL}"
         echo "INFO    [qq metacentrum installer] Installation completed on ${HOST}"
     else
         echo "WARN    [qq metacentrum installer] Could not connect to ${HOST}. Skipping."